chore(server)!: move enhancers to a dedicated directory
