+++ routes/enhancer/admin/resource_config/departments.get.ts
