+++ routes/enhancer/admin/resource_config/index.get.ts
