+++ routes/enhancer/admin/resource_config/router.ts
@@ -1,0 +1,18 @@
+import Router from "!/bases/router"
+import GetIndex from "!%/enhancer/admin/resource_config/index.get"
+import GetUsersRoute from "!%/enhancer/admin/resource_config/users.get"
+import GetRolesRoute from "!%/enhancer/admin/resource_config/roles.get"
+import GetDepartmentsRoute from "!%/enhancer/admin/resource_config/departments.get"
+
+export default class extends Router {
+	constructor() {
+		super()
+
+		this.useControllers([
+			new GetIndex(),
+			new GetUsersRoute(),
+			new GetRolesRoute(),
+			new GetDepartmentsRoute()
+		])
+	}
+}
