+++ routes/enhancer/admin/router.ts
@@ -1,0 +1,12 @@
+import Router from "!/bases/router"
+import ResourceConfigurationRouter from "!%/enhancer/admin/resource_config/router"
+
+export default class extends Router {
+	constructor() {
+		super()
+
+		this.useRouters([
+			new ResourceConfigurationRouter()
+		])
+	}
+}
