+++ routes/enhancer/chat/room(uuid).get.ts
