+++ routes/enhancer/chat/router.ts
@@ -1,12 +1,14 @@
 import Router from "!/bases/router"
-import GetCreateRoute from "!%/department/create.get"
+import GetJoin from "!%/enhancer/chat/join.get"
+import GetRoom from "!%/enhancer/chat/room(uuid).get"
 
 export default class extends Router {
 	constructor() {
 		super()
 
 		this.useControllers([
-			new GetCreateRoute()
+			new GetJoin(),
+			new GetRoom()
 		])
 	}
 }
