+++ routes/enhancer/department/router.ts
@@ -1,14 +1,12 @@
 import Router from "!/bases/router"
-import GetJoin from "!%/chat/join.get"
-import GetRoom from "!%/chat/room(uuid).get"
+import GetCreateRoute from "!%/enhancer/department/create.get"
 
 export default class extends Router {
 	constructor() {
 		super()
 
 		this.useControllers([
-			new GetJoin(),
-			new GetRoom()
+			new GetCreateRoute()
 		])
 	}
 }
