+++ routes/enhancer/settings/router.ts
@@ -1,13 +1,12 @@
 import Router from "!/bases/router"
-import GetIndex from "!%/settings/index.get"
+import GetIndex from "!%/enhancer/settings/index.get"
 
 export default class extends Router {
-
 	constructor() {
 		super()
 
 		this.useControllers([
-			new GetIndex(),
+			new GetIndex()
 		])
 	}
 }
